requirements.txt
pandas>=2.0.0
lxml>=5.0.0
beautifulsoup4>=4.12.0
html5lib>=1.1
jinja2>=3.1.0
ebooklib>=0.18
odfpy>=1.4.1
reportlab>=4.0.0
python-pptx>=0.6.0
python-docx>=1.1.0
openpyxl>=3.1.0
xlrd>=2.0.1
Pillow>=10.0.0
rawpy>=0.18.0
pillow-heif>=0.13.0
markdown>=3.5.0
chardet>=5.0.0
pyarrow>=14.0.0
